refactor: Simplify main page components path- Remove directories which has only one component.- Create index file in directory that has more than one component
