+++ client/src/components/main/Category.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import ProductCategory from './category/ProductCategory';
+import ProductCategory from './ProductCategory';
 
 const CategoryBlock = styled.div`
   margin-bottom: 47px;
