+++ client/src/components/main/ProductEssential.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import styled from 'styled-components';
-import ProductInfo from '../common/ProductInfo';
-import Refresh from '../common/Refresh';
-import { Query } from 'react-apollo';
-import getRandomInt from '../../../utils/random';
-import { GET_PRODUCT_SIMPLE } from '../main-query';
+import ProductInfo from './common/ProductInfo';
+import Refresh from './common/Refresh';
 
-const ProductWhatEatBlock = styled.div`
+import { Query } from 'react-apollo';
+import getRandomInt from '../../utils/random';
+import { GET_PRODUCT_SIMPLE } from './main-query';
+
+const ProductEssentialBlock = styled.div`
   .ProductTitle {
     padding: 1rem;
     background-color: white;
@@ -14,11 +15,11 @@ const ProductWhatEatBlock = styled.div`
     font-weight: bold;
   }
   .ProductInfo {
-    padding-left: 0.5rem;
-    padding-right: 0.5rem;
     background-color: white;
     display: flex;
     flex-wrap: wrap;
+    padding-left: 0.5rem;
+    padding-right: 0.5rem;
     justify-content: space-between;
     height: auto;
     width: auto;
@@ -33,21 +34,24 @@ const ProductWhatEatBlock = styled.div`
     padding-bottom: 0.3rem;
   }
 `;
-const cursor = getRandomInt(450, 500);
 
-function ProductWhatEat() {
+const cursor = getRandomInt(6546, 6607);
+
+function ProductEssential() {
   return (
-    <ProductWhatEatBlock>
-      <div className="ProductTitle">지금 뭐 먹지?</div>
+    <ProductEssentialBlock>
+      <div className="ProductTitle">지금 필요한 생필품!</div>
       <div className="ProductInfo">
-      <Query query={GET_PRODUCT_SIMPLE} variables={{ take: 9, cursor:cursor}}>
+        <Query
+          query={GET_PRODUCT_SIMPLE}
+          variables={{ take: 9, cursor: cursor }}>
           {({ data, loading, error }) => {
             if (loading || error) return '';
             return data.products.products.map((product, idx) => {
               return (
                 <ProductInfo
-                  id={product.id}
                   key={idx}
+                  id={product.id}
                   title={product.name}
                   price={product.price}
                   url={product.img_url}></ProductInfo>
@@ -57,10 +61,10 @@ function ProductWhatEat() {
         </Query>
       </div>
       <div className="Refresh">
-        <Refresh className="Refresh" title={'지금 뭐 먹지? '}></Refresh>
+        <Refresh title={'지금 필요한 생필품! '}></Refresh>
       </div>
-    </ProductWhatEatBlock>
+    </ProductEssentialBlock>
   );
 }
 
-export default ProductWhatEat;
+export default ProductEssential;
