+++ client/src/components/main/ProductFlashDiscount/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProductFlashDiscount';
